requirements.txt
asgiref==3.8.1
Django==5.0.4
psycopg2==2.9.9
sqlparse==0.5.0
pillow~=10.3.0
